refactor: Move Document to own file
